+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2026-04-29
 ### Added
 - Rename the selected Assets by enumerate them
 
@@ -29,7 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/ToolSystem/compare/1.2.1...main
+[Unreleased]: https://github.com/1mbitshorde/ToolSystem/compare/1.3.0...main
+[1.3.0]: https://github.com/1mbitshorde/ToolSystem/tree/1.3.0/
 [1.2.1]: https://github.com/1mbitshorde/ToolSystem/tree/1.2.1/
 [1.2.0]: https://github.com/1mbitshorde/ToolSystem/tree/1.2.0/
 [1.1.0]: https://github.com/1mbitshorde/ToolSystem/tree/1.1.0/